Migrated to GTK+3 (PyGObject)

+++ wasteland.py
@@ -1,16 +1,29 @@
 #!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Waste land, Clean up unused thumbnails
+#
+# Copyright (C) 2006-2012 Germán Poo-Caamaño <gpoo@gnome.org>
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, write to the Free Software Foundation, Inc.,
+# 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-from gnomevfs import get_local_path_from_uri
 from gc import collect
+import os
 import os.path
 
-import gtk
-import gtk.gdk
-import gnome.ui
-import gobject
-import os
-import gtk.glade
-import pango
+from gi.repository import Gtk, GObject, Pango, Gio, GdkPixbuf
 
 def bytes_to_string(size):
     """Return a number in a easier way to read"""
@@ -32,45 +45,44 @@ def bytes_to_string(size):
 
 
 class ThumbnailChecker:
-    OK = 0
-    EXTERNAL = 1
-    ORPHAN = 2
-    NON_FD = 3
-    INVALID = 4
+    OK, EXTERNAL, ORPHAN, NON_FD, INVALID = range(0, 5)
 
     def __init__(self):
         self.id = None
         self.first_time = True
 
-        xml = gtk.glade.XML('thumbnail-checker.glade', None, None)
-        self.model = gtk.TreeStore(str, str, str)
+        builder = Gtk.Builder()
+        builder.add_from_file(os.path.join(os.path.dirname(__file__),
+                                       'wasteland.glade'))
+        self.model = Gtk.TreeStore(str, str, str)
 
-        self.window = xml.get_widget('window')
-        self.treeview = xml.get_widget('treeview')
-        self.progressbar = xml.get_widget('progressbar')
-        self.progress = xml.get_widget('progress')
-        self.button_delete = xml.get_widget('button_delete')
-        self.button_start = xml.get_widget('button_start')
-        self.button_stop = xml.get_widget('button_stop')
-        xml.signal_autoconnect(self)
+        self.window = builder.get_object('window')
+        self.treeview = builder.get_object('treeview')
+        self.progressbar = builder.get_object('progressbar')
+        self.progress = builder.get_object('progress')
+        self.progress_uri = builder.get_object('uri')
+        self.button_delete = builder.get_object('button_delete')
+        self.button_start = builder.get_object('button_start')
+        self.button_stop = builder.get_object('button_stop')
+        builder.connect_signals(self)
 
         self.treeview.set_search_column(1)
         self.treeview.set_model(self.model)
         selection = self.treeview.get_selection()
-        selection.set_mode(gtk.SELECTION_MULTIPLE)
+        selection.set_mode(Gtk.SelectionMode.MULTIPLE)
         selection.connect("changed", self.on_selection_changed)
 
-        renderer = gtk.CellRendererText()
-        renderer.set_property("ellipsize", pango.ELLIPSIZE_MIDDLE)
-        column = gtk.TreeViewColumn("File name or Thumbnail file",
+        renderer = Gtk.CellRendererText()
+        renderer.set_property("ellipsize", Pango.EllipsizeMode.MIDDLE)
+        column = Gtk.TreeViewColumn("File name or Thumbnail file",
                                     renderer, text=0)
         column.set_resizable(True)
         column.set_expand(True)
         self.treeview.append_column(column)
 
-        renderer = gtk.CellRendererText()
+        renderer = Gtk.CellRendererText()
         renderer.set_property('xalign', 1.0)
-        column = gtk.TreeViewColumn("Size", renderer, text=1)
+        column = Gtk.TreeViewColumn("Size", renderer, text=1)
         column.set_resizable(True)
         column.set_min_width(80)
         column.set_resizable(True)
@@ -80,24 +92,25 @@ class ThumbnailChecker:
         self.window.show_all()
 
     def on_quit(self, *args):
-        gtk.main_quit()
+        Gtk.main_quit()
 
     def on_button_delete_clicked(self, button, *args):
         selection = self.treeview.get_selection()
         model, paths = selection.get_selected_rows()
         refs = []
         for path in paths:
-            refs.append(gtk.TreeRowReference(model, path))
+            refs.append(Gtk.TreeRowReference.new(model, path))
 
         for ref in refs:
             path = ref.get_path()
             iter = model.get_iter(path)
             filename = model.get(iter, 2)[0]
-            os.unlink(os.path.expanduser(filename))
-            model.remove(iter)
+            if filename is not None:
+                os.unlink(os.path.expanduser(filename))
+                model.remove(iter)
 
     def on_button_stop_clicked(self, button, *args):
-        gobject.source_remove(self.id)
+        GObject.source_remove(self.id)
         self.button_stop.set_sensitive(False)
         self.button_start.set_sensitive(True)
 
@@ -107,7 +120,7 @@ class ThumbnailChecker:
         if self.first_time:
             self.first_time = False
             self.task = self.walk()
-        self.id = gobject.idle_add(self.task.next)
+        self.id = GObject.idle_add(self.task.next)
 
     def walk(self, *args):
         (self.invalid_size, self.invalid_count) = (0, 0)
@@ -130,7 +143,7 @@ class ThumbnailChecker:
 
             text = "%d of %d" % (0, len(files))
             self.progressbar.set_text(text)
-            text = "Processing %s ..." % root
+            text = "Processing %s:" % (os.path.basename(root))
             self.progress.set_text(text)
 
             for name in files:
@@ -139,9 +152,13 @@ class ThumbnailChecker:
                 filename = os.path.join(root, name)
 
                 self.progressbar.set_fraction(i / len(files))
+                text = "%d of %d" % (i, len(files))
                 self.progressbar.set_text(text)
 
-                (result, resource) = self.verify_thumbnail(filename)
+                (result, resource, uri) = self.verify_thumbnail(filename)
+
+                self.progress_uri.set_text(uri)
+
                 if resource.startswith(homedir):
                     resource = resource.replace(homedir, '~')
 
@@ -168,13 +185,11 @@ class ThumbnailChecker:
         self.walk_done()
         yield False
 
-    #def delete_selected_thumbnail(self, treemodel, path, iter):
-    #    file = treemodel.get(iter, 2)[0]
-    #    os.unlink(os.path.expanduser(file))
     def walk_done(self):
-        gobject.source_remove(self.id)
+        GObject.source_remove(self.id)
         self.progressbar.set_text('Done')
         self.progress.set_text('')
+        self.progress_uri.set_text('')
         self.button_stop.set_sensitive(False)
 
     def on_selection_changed(self, selection):
@@ -183,31 +198,28 @@ class ThumbnailChecker:
 
     def verify_thumbnail(self, filename):
         try:
-            pixbuf = gtk.gdk.pixbuf_new_from_file(filename)
+            pixbuf = GdkPixbuf.Pixbuf.new_from_file(filename)
             uri = pixbuf.get_option('tEXt::Thumb::URI')
         except:
             # Broken thumbnail
             uri = ''
 
-        try:
-            path = get_local_path_from_uri(uri)
-        except:
-            # External resource or invalid uri
-            path = ''
+        gfile = Gio.File.new_for_uri(uri)
+        path = gfile.get_path()
 
-        if len(path) == 0 and uri is not None and len(uri):
-            return (self.EXTERNAL, uri)
-        if len(path) and not os.path.lexists(path):
-            return (self.ORPHAN, filename)
+        if path is None and uri is not None and len(uri):
+            return (self.EXTERNAL, uri, uri)
+        if path and not os.path.lexists(path):
+            return (self.ORPHAN, filename, uri)
         if uri is None:
-            return (self.NON_FD, filename)
+            return (self.NON_FD, filename, uri)
         if len(uri) == 0:
-            return (self.INVALID, filename)
-        return (self.OK, filename)
+            return (self.INVALID, filename, uri)
+        return (self.OK, filename, uri)
 
     def compute(self, filename, item, element, text):
         sum_size = getattr(self, item + '_size')
-        counter  = getattr(self, item + '_count')
+        counter = getattr(self, item + '_count')
         iter = getattr(self, item + '_iter')
 
         size = os.path.getsize(filename)
@@ -224,8 +236,9 @@ class ThumbnailChecker:
         self.model.set(iter, 1, str_size)
         self.model.set(iter, 0, text % counter)
 
+
 if __name__ == "__main__":
     checker = ThumbnailChecker()
     checker.show()
 
-    gtk.main()
+    Gtk.main()
